Put commands in a properly injectable manner
