+++ bot/src/main/kotlin/org/epilink/bot/LinkServerEnvironment.kt
@@ -17,6 +17,7 @@ import org.epilink.bot.config.LinkConfiguration
 import org.epilink.bot.db.*
 import org.epilink.bot.db.exposed.SQLiteExposedFacadeImpl
 import org.epilink.bot.discord.*
+import org.epilink.bot.discord.cmd.UpdateCommand
 import org.epilink.bot.http.*
 import org.epilink.bot.http.endpoints.*
 import org.epilink.bot.rulebook.Rulebook
@@ -87,6 +88,7 @@ class LinkServerEnvironment(
         single<LinkBanManager> { LinkBanManagerImpl() }
         single<LinkDiscordCommands> { LinkDiscordCommandsImpl() }
         single<LinkDiscordTargets> { LinkDiscordTargetsImpl() }
+        single<List<Command>>(named("discord.commands")) { listOf(UpdateCommand()) }
     }
 
     /**
